+++ src/main/java/org/jsoup/nodes/DataNode.java
@@ -1,6 +1,8 @@
 package org.jsoup.nodes;
 
 import java.io.IOException;
+import java.util.regex.Pattern;
+
 import org.jsoup.nodes.Entities.EscapeMode;
 
 /**
@@ -41,14 +43,18 @@ public class DataNode extends LeafNode {
 
     @Override
     void outerHtmlHead(Appendable accum, int depth, Document.OutputSettings out) throws IOException {
-        if (out.syntax() == Document.OutputSettings.Syntax.xml) {
-            // In XML mode, output data nodes as CDATA, so can parse as XML
-            accum
-                .append("<![CDATA[")
-                .append(getWholeData())
-                .append("]]>");
+        /* For XML output, escape the DataNode in a CData section. The data may contain pseudo-CData content if it was
+        parsed as HTML, so don't double up Cdata. Output in polygot HTML / XHTML / XML format. */
+        final String data = getWholeData();
+        if (out.syntax() == Document.OutputSettings.Syntax.xml && !data.contains("<![CDATA[")) {
+            if (hasParent() && parentNode.normalName().equals("script"))
+                accum.append("//<![CDATA[\n").append(data).append("\n//]]>");
+            else if (hasParent() && parentNode.normalName().equals("style"))
+                accum.append("/*<![CDATA[*/\n").append(data).append("\n/*]]>*/");
+            else
+                accum.append("<![CDATA[").append(data).append("]]>");
         } else {
-            // In HTML, data is not escaped in return from data nodes, so " in script, style is plain
+            // In HTML, data is not escaped in the output of data nodes, so < and & in script, style is OK
             accum.append(getWholeData());
         }
     }